backend/requirements.txt
flask==3.0.2
pdfplumber==0.11.0
python-docx==1.1.0
rapidfuzz==3.9.3
pytest==8.1.1
fpdf2==2.7.9
requests==2.31.0
beautifulsoup4==4.12.3
